requirements.txt
Jinja2==3.1.6
openai==1.109.0
pandas==2.3.3
pip-check-reqs==2.5.5
pipreqs==0.4.13
pytest==8.4.2
python-dotenv==1.1.1
ruff==0.13.1
tabulate==0.9.0
